Refactor exporting/importing sap system details component
