+++ assets/js/components/DatabaseDetails/DatabaseDetails.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useParams } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import { getDatabaseDetail } from '@state/selectors';
-import GenericSystemDetails from '@components/SapSystemDetails/GenericSystemDetail';
+import { GenericSystemDetails } from '@components/SapSystemDetails';
 import { DATABASE_TYPE } from '@lib/model';
 
 const DatabaseDetails = () => {
