+++ assets/js/components/SapSystemDetails/index.js
@@ -1,3 +1,9 @@
 import SapSystemDetails from './SapSystemDetails';
 
+export {
+  GenericSystemDetails,
+  Features,
+  InstanceStatus,
+} from './GenericSystemDetails';
+
 export default SapSystemDetails;
